app/requirements.txt
fastapi==0.100.0
python-dotenv==1.0.0
